FEAT: impelment jwt auth logic and add some todos for further improvement
