+++ API/src/main/java/com/litcode/server/security/JwtTokenProvider.java
@@ -9,20 +9,17 @@ import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
 import io.jsonwebtoken.security.SecurityException;
 import lombok.extern.slf4j.Slf4j;
-
-import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 
 import java.security.Key;
 import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
-
 import javax.annotation.PostConstruct;
 
 @Slf4j
 @Component
 public class JwtTokenProvider {
+	// TODO
+	// env var
 	private String jwtSecret = "dasdasddsadasdasdasdasgnfsdkgsgsfdklgnfdnmekehwdkfhesdfhsfhksdhfksdfnv";
 	private Key jwtKey;
 	private int jwtExpirationInMs = 3600000;
@@ -34,15 +31,13 @@ public class JwtTokenProvider {
 		this.jwtKey = Keys.hmacShaKeyFor(keyBytes);
 	}
 
-	public String generateToken(Authentication authentication) {
-		UserPrincipal userPrincipal = (UserPrincipal) authentication.getPrincipal();
-
+	public String generateToken(String userId) {
 		Date now = new Date();
 		Date expiryDate = new Date(now.getTime() + jwtExpirationInMs);
 
 		return Jwts.builder()
 				.setIssuedAt(new Date())
-				.claim("userId", userPrincipal.getUsername())
+				.claim("userId", userId)
 				.setExpiration(expiryDate)
 				.signWith(jwtKey)
 				.compact();
@@ -57,29 +52,25 @@ public class JwtTokenProvider {
 				.getBody();
 
 		return claims.get("userId").toString();
-
 	}
 
 	public boolean validateToken(String authToken) {
 
+		// TODO
+		// error message handling according to validation error
 		try {
 			Jwts.parserBuilder().setSigningKey(jwtKey).build().parseClaimsJws(authToken);
 			return true;
 		} catch (SecurityException ex) {
-			// LOGGER.error("Invalid JWT signature");
-			System.out.println("invalid jwt sig");
+			log.error("invalid jwt sig");
 		} catch (MalformedJwtException ex) {
-			// LOGGER.error("Invalid JWT token");
-			System.out.println("invalid jwt tok");
-
+			log.error("invalid jwt tok");
 		} catch (ExpiredJwtException ex) {
-			// LOGGER.error("Expired JWT token");
-			System.out.println("invalid jwt exp");
+			log.error("invalid jwt exp");
 		} catch (UnsupportedJwtException ex) {
-			System.out.println("invalid jwt sup");
+			log.error("invalid jwt sup");
 		} catch (IllegalArgumentException ex) {
-			System.out.println("invalid jwt empt");
-			// LOGGER.error("JWT claims string is empty");
+			log.error("invalid jwt empt");
 		}
 		return false;
 	}
