--- API/src/main/java/com/litcode/server/security/UserPrincipal.java
@@ -1,56 +1,0 @@
-package com.litcode.server.security;
-
-import java.util.Collection;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import com.litcode.server.model.UserModel;
-
-public class UserPrincipal implements UserDetails {
-
-	private UserModel userModel;
-
-	public UserPrincipal(UserModel userModel) {
-		this.userModel = userModel;
-	}
-
-	@Override
-	public boolean isEnabled() {
-		return true;
-	}
-
-	@Override
-	public Collection<? extends GrantedAuthority> getAuthorities() {
-		return null;
-	}
-
-	@Override
-	public String getPassword() {
-		return userModel.getPassword();
-	}
-
-	@Override
-	public String getUsername() {
-		return userModel.getUserId();
-	}
-
-	public String getUserEmail() {
-		return userModel.getEmail();
-	}
-
-	@Override
-	public boolean isAccountNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isAccountNonLocked() {
-		return true;
-	}
-
-	@Override
-	public boolean isCredentialsNonExpired() {
-		return true;
-	}
-}
